stop running the re-create loop on menu exit if config disabled
Currently, if the config is disabled, we try to convert everything back
to normal size on closing the carpenter menu.

This is unnecessary. Originally I planned for this to be how we fixed
the ponds if configuration was changed. However, a future change will
introduce a proper config saving function that handles this.

Stop trying to re-create ponds when the config is disabled, except when
saving the file.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>