factor out RecreateAllPonds function
We perform almost the exact same logic for converting ponds in 3
locations, when the carpenter menu closes, when we save, and when we
load.

All of these flows need to find all the buildings and check for fish
ponds which are the wrong type. In the case of closing the carpenter
menu or loading, we're converting from the default pond type to the
smaller type. On saving, we convert from the small type to the normal
type.

A future fix is going to add support for multiple game locations, which
requires modifying this logic. Instead of modifying it in three places,
refactor to use a common funciton.

Introduce RecreateAllPonds which will loop over the buildings and locate
the ponds. It takes a boolean argument which indicates whether we're
converting to the small size or the normal size.

If we're converting to small size, ignore buildings which are already
small. If we're converting to the normal size, ignore buildings which
are already big.

The new function uses a standard foreach loop in order to better prepare
for handling multiple game locations.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>