update all ponds on configuration change
Now that we have RecreateAllPonds, update the config menu so that it
will automatically convert ponds on saving. If the config is disabled,
convert everything to the normal size. If the config is enabled, convert
everything to the small size.

This should allow dynamically disabling the mod without needing to save
the game to see the change.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>