Use BuildingListChanged instead of MenuChanged event
Any time a new pond is added, we need to re-create it as a smaller pond.
Currently this is achieved by using the MenuChanged event to check when
the carpenter menu is closed.

This is problematic because it means if a mod adds a new menu for
constructing buildings, we do not properly trigger, and leave the
building in the old state.

The game has a much better event to handle this, the
World.BuildingListChanged. This event is better for a few reasons:

a) it triggers whenever a building is added for any reason, whether from
a construction menu or not
b) it *only* triggers when buildings are added or removed, meaning we
don't waste time checking for buildings when unnecessary.
c) it includes the exact list of new buildings, meaning we don't need to
scan every building over again

Factor out a RecreatePondsAt() function which takes a location, the set
of buildings to check, and whether we're passing the smallSize or not.

Replace the OnMenuChanged event with OnBuildingListChanged, and use the
newly created RecreatePondsAt() function. This ensures we only check new
buildings, and avoid needing to check after every single menu close.

Additionally, handle the LocationListChanged event. For any new location
added, go ahead and also check those buildings. This is (probably) not
strictly necessary since we handle changing all locations on save and
load for compatibility reasons. However, if a location *is* dynamically
added during play for some reason, we should check and handle this.

This should fix the ponds so that they work in all locations regardless
of how they got added.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>