add option to keep the smaller size on saving
Several users reported issues with the ponds destroying nearby crops
when saving. This is a result of restoring the pond to its 5x5 size.

This was done in order to avoid issues with removing the mod. However,
the visual issues with smaller ponds are less of an issue than
destroying crops.

Add a new option to keep the smaller size when saving, and make it the
default. If you want to migrate away from the mod, you should disable
this option and then save the game.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>